requirements.txt
requests>=2.25.1
dnspython>=2.0.0
tqdm>=4.60.0
